ç fonctionne parfait en http avec recul du status code il manque plus qu'a comprendre comment pin un cert

+++ MSW - iPhone/MSW2/ApiMethod.h
@@ -11,7 +11,7 @@
 #import "AFNetworking.h"
 
 NSString *Id_global;
-int       code_global;
+long      code_global;
 
 @interface ApiMethod : NSObject
